+++ contrib/adapters/objectstores3/doc.go
@@ -1,0 +1,2 @@
+// Package objectstores3 provides an experimental S3-compatible object storage adapter.
+package objectstores3